backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv
sqlalchemy
PyJWT
Authlib
httpx
itsdangerous
bcrypt==3.2.2
passlib[bcrypt]
psycopg2-binary
python-jose[cryptography]
python-multipart
google-auth
pandas